Add package.toml
Add a TOML version of package.json. Ideally, configurations should all be authored in TOML, not JSON, which is more suitable for data exchange between machines.

The TOML version has no function right now, but rather serves as a documentation to be reflected upon when we explore full interoperability (comment preservation, please!) between TOML and JSON in the future.